Updating API reference from int to decimal from 22.3 changes

+++ src/EncompassRest.Loans.v1/Loans/v1/Employment.cs
@@ -218,7 +218,7 @@ public sealed partial class Employment : DirtyExtensibleObject, IIdentifiable
     /// <summary>
     /// Employment MonthlyIncomeAmount
     /// </summary>
-    public int? MonthlyIncomeAmount { get => GetValue<int?>(); set => SetValue(value); }
+    public decimal? MonthlyIncomeAmount { get => GetValue<decimal?>(); set => SetValue(value); }
 
     /// <summary>
     /// Employment NoLinkToDocTrackIndicator [BENN97], [CENN97]
